New logic for resolving schedule
Due to the current scheduling logic being non-functional because of
deprecated resources on NFL.com, this logic has been updated in the
following way:

* The season, phase and week is simply calculated using the NFL's
scheduling rules (it all revolves around Labor day, which is the first
monday in september every year)
* This means that alternation between regular- and postseason
schedules is no longer required
* This also means that no external calls are made to determine where we
are at in a season

Please see the code documentation for further explanation. Unit tests
are provided, testing the whole of the 2018 season as well as critical
points (going from pre to regular season etc) from 2014 an onward to see
that the logic works over time, handles leap years etc.