Enhanced scheduling logic (#23)
The regular, no-arg- scheduling updating logic now works like this:

* Check which weeks we are _missing_ in `schedule.json` for the current season, if any. The missing weeks will be chronologically ordered.
* If the current week is among the missing weeks, move it to the front of the list.
* If the current week is not among the missing weeks, add the current week to the front of the list of missing weeks (which may be empty).
* Try to get the weeks, one by one in order. Abort on failure.

Also added a print statement to a) make it clearer to the user that this
happens (if the schedule hasn't been updated in a day) when simply
importing the module and b) to know which weeks were updated, especially
if more than the current week.

Unit tests are provided for the functions that

* Check for missing weeks
* Orders those weeks to make sure that the current week is always the first to try and get from NFL.com

There are however, other changes that are made to the code which are *not* unit tested. This is because of the difficulties to test code that is previously untested, makes calls over several files and maybe most importantly, makes external calls. These external calls would have to be mocked and at this time I couldn't figure out a nice, clean way to do this without major refactoring which is outside the scope of issue #18. Some of it is probably better to integration test rather than unit test.

Those parts of the code are manually tested, though.